+++ Runtime/Bindings/_Builders/Experimental/RepeatBindingBuilder.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using UnityEngine;
 using VentiCola.UI.Bindings.LowLevel;
 using VentiCola.UI.Bindings.LowLevel.Experimental;
@@ -7,11 +8,15 @@ namespace VentiCola.UI.Bindings.Experimental
 {
     public static class RepeatBindingBuilder
     {
-        public static GameObject RepeatForEachOf<T>(this GameObject self, Func<GameObject, ReactiveList<T>> collection, Action<int, T> renderItemAction = null)
+        public static GameObject RepeatForEachOf<T>(
+            this GameObject self,
+            Func<GameObject, ReactiveList<T>> collection,
+            Action<ForEachItem<T>> renderItemAction = null,
+            IEqualityComparer<T> customEqualityComparer = null)
         {
             BaseBinding
                 .Allocate<ForEachBinding<T>>()
-                .Initalize(self, collection, renderItemAction);
+                .Initalize(self, collection, renderItemAction, customEqualityComparer);
             return self;
         }